.agents/skills/criticar-url/scripts/pyproject.toml
[project]
name = "criticar-url-skill"
version = "0.1.0"
description = "Ambiente Python local da skill criticar-url"
requires-python = ">=3.12"
dependencies = [
  "pyyaml>=6.0",
]

[tool.uv]
package = false
